+++ nucliadb/tests/ndbfixtures/standalone.py
@@ -20,7 +20,6 @@
 import base64
 import logging
 import os
-import uuid
 from pathlib import Path
 from unittest.mock import patch
 
@@ -30,7 +29,7 @@ from httpx import AsyncClient
 
 from nucliadb.common.maindb.exceptions import UnsetUtility
 from nucliadb.common.maindb.utils import get_driver
-from nucliadb.search.api.v1.router import KB_PREFIX, KBS_PREFIX
+from nucliadb.ingest.settings import DriverSettings
 from nucliadb.standalone.config import config_nucliadb
 from nucliadb.standalone.run import run_async_nucliadb
 from nucliadb.standalone.settings import Settings
@@ -58,17 +57,61 @@ logger = logging.getLogger(__name__)
 
 
 @pytest.fixture(scope="function")
-async def knowledgebox_one(nucliadb_writer_manager: AsyncClient):
-    kbslug = str(uuid.uuid4())
-    data = {"slug": kbslug}
-    resp = await nucliadb_writer_manager.post(f"/{KBS_PREFIX}", json=data)
-    assert resp.status_code == 201
-    kbid = resp.json()["uuid"]
+async def nucliadb(
+    endecryptor_settings,
+    dummy_processing,
+    analytics_disabled,
+    maindb_settings: DriverSettings,
+    storage: Storage,
+    storage_settings,
+    tmp_path: Path,
+    learning_config,
+):
+    from nucliadb.common.cluster import manager
 
-    yield kbid
+    manager.INDEX_NODES.clear()
 
-    resp = await nucliadb_writer_manager.delete(f"/{KB_PREFIX}/{kbid}")
-    assert resp.status_code == 200
+    # we need to force DATA_PATH updates to run every test on the proper
+    # temporary directory
+    data_path = str((tmp_path / "node").absolute())
+    with patch.dict(os.environ, {"DATA_PATH": data_path}, clear=False):
+        settings = Settings(
+            data_path=data_path,
+            http_port=free_port(),
+            ingest_grpc_port=free_port(),
+            train_grpc_port=free_port(),
+            standalone_node_port=free_port(),
+            log_format_type=LogFormatType.PLAIN,
+            log_output_type=LogOutputType.FILE,
+            **maindb_settings.model_dump(),
+            **storage_settings,
+        )
+
+        config_nucliadb(settings)
+
+        # Make sure tests don't write logs outside of the tmpdir
+        os.environ["ERROR_LOG"] = str((tmp_path / "logs" / "error.log").absolute())
+        os.environ["ACCESS_LOG"] = str((tmp_path / "logs" / "access.log").absolute())
+        os.environ["INFO_LOG"] = str((tmp_path / "logs" / "info.log").absolute())
+
+        setup_logging(
+            settings=LogSettings(
+                log_output_type=LogOutputType.FILE,
+                log_format_type=LogFormatType.PLAIN,
+                debug=False,
+                log_level=LogLevel.WARNING,
+            )
+        )
+        server = await run_async_nucliadb(settings)
+        assert server.started, "Nucliadb server did not start correctly"
+
+        yield settings
+
+        await maybe_cleanup_maindb()
+
+        reset_config()
+        clear_global_cache()
+        await server.shutdown()
 
 
 @pytest.fixture(scope="function", autouse=True)
@@ -94,62 +137,7 @@ def endecryptor_settings():
         yield
 
 
-@pytest.fixture(scope="function")
-async def nucliadb(
-    endecryptor_settings,
-    dummy_processing,
-    analytics_disabled,
-    maindb_settings,
-    storage: Storage,
-    storage_settings,
-    tmp_path: Path,
-    learning_config,
-):
-    from nucliadb.common.cluster import manager
-
-    manager.INDEX_NODES.clear()
-
-    # we need to force DATA_PATH updates to run every test on the proper
-    # temporary directory
-    data_path = str((tmp_path / "node").absolute())
-    os.environ["DATA_PATH"] = data_path
-    settings = Settings(
-        data_path=data_path,
-        http_port=free_port(),
-        ingest_grpc_port=free_port(),
-        train_grpc_port=free_port(),
-        standalone_node_port=free_port(),
-        log_format_type=LogFormatType.PLAIN,
-        log_output_type=LogOutputType.FILE,
-        **maindb_settings.model_dump(),
-        **storage_settings,
-    )
-
-    config_nucliadb(settings)
-
-    # Make sure tests don't write logs outside of the tmpdir
-    os.environ["ERROR_LOG"] = str((tmp_path / "logs" / "error.log").absolute())
-    os.environ["ACCESS_LOG"] = str((tmp_path / "logs" / "access.log").absolute())
-    os.environ["INFO_LOG"] = str((tmp_path / "logs" / "info.log").absolute())
-
-    setup_logging(
-        settings=LogSettings(
-            log_output_type=LogOutputType.FILE,
-            log_format_type=LogFormatType.PLAIN,
-            debug=False,
-            log_level=LogLevel.WARNING,
-        )
-    )
-    server = await run_async_nucliadb(settings)
-    assert server.started, "Nucliadb server did not start correctly"
-
-    yield settings
-
-    await maybe_cleanup_maindb()
-
-    reset_config()
-    clear_global_cache()
-    await server.shutdown()
+# Derived
 
 
 @pytest.fixture(scope="function")
@@ -173,6 +161,9 @@ async def nucliadb_grpc(nucliadb: Settings):
 async def nucliadb_train(nucliadb: Settings):
     stub = TrainStub(aio.insecure_channel(f"localhost:{nucliadb.train_grpc_port}"))
     return stub
+
+
+# Utils
 
 
 async def maybe_cleanup_maindb():
